refactor(web-app): reorganize Stellar helpers and shared tokenUtils
- Move stellar-network, stellar-wallet-kit, sign-stellar-transaction into lib/helpers/stellar/
- Group wallet logic in stellar/wallet/ (walletKit, signTransaction, balances, index)
- Move tokenUtils from stellar/ to lib/helpers/ as shared Stellar+EVM util
- Update all imports across features, hooks, providers, and services

Made-with: Cursor
