+++ apps/web-app/src/hooks/useStellarWallet.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { ISupportedWallet } from "@creit.tech/stellar-wallets-kit";
-import { getStellarWalletKit } from "@/lib/helpers/stellar-wallet-kit";
+import { getStellarWalletKit } from "@/lib/helpers/stellar/wallet";
 import { useStellarWalletStore } from "@/stores/stellarWalletStore";
 
 export function useStellarWallet() {
