+++ apps/web-app/src/hooks/useTokenBalance.ts
@@ -13,7 +13,7 @@ import {
   horizonUrl,
 } from "@/lib/config/stellar.config";
 import { useWallet } from "./useWallet";
-import { fromSmallestUnit } from "@/lib/helpers/stellar/tokenUtils";
+import { fromSmallestUnit } from "@/lib/helpers/tokenUtils";
 import { getTokens, getAvailableTokens } from "@/lib/helpers/stellar/soroswap";
 import { getTokenAddress } from "@/lib/helpers/stellar/soroswap";
