+++ apps/web-app/src/hooks/useWallet.ts
@@ -1,9 +1,11 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import type { MappedBalances } from "@/lib/helpers/stellar/wallet";
-import { fetchBalances } from "@/lib/helpers/stellar/wallet";
-import { signStellarTransactionWithWallet } from "@/lib/helpers/sign-stellar-transaction";
+import {
+  fetchBalances,
+  signStellarTransactionWithWallet,
+  type MappedBalances,
+} from "@/lib/helpers/stellar/wallet";
 import { stellarNetwork, networkPassphrase } from "@/lib/constants/network";
 import { useStellarWalletStore } from "@/stores/stellarWalletStore";
