+++ apps/web-app/src/lib/helpers/stellar/lending.ts
@@ -18,7 +18,7 @@ import {
   horizonUrl,
   stellarNetwork,
 } from "@/lib/constants/network";
-import { toSmallestUnit } from "./tokenUtils";
+import { toSmallestUnit } from "../tokenUtils";
 import { extractContractError } from "./contractErrors";
 
 /**
