+++ apps/web-app/src/lib/helpers/stellar/network.ts
